Add k8s custom policy tag handler for test
